scripts/requirements.txt
gpiozero>=1.5.1
pn532pi>=1.3
python-mpd2>=3.0.1
